Centralize default properties for nebula tasks
