+++ src/main/kotlin/nebula/plugin/bintray/NebulaBintrayVersionTask.kt
@@ -1,37 +1,9 @@
 package nebula.plugin.bintray
 
-import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
-import org.gradle.api.provider.Property
-import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.TaskAction
-import org.gradle.kotlin.dsl.property
 
-open class NebulaBintrayVersionTask : DefaultTask() {
-    @Input
-    val user: Property<String> = project.objects.property()
-    @Input
-    val apiKey: Property<String> = project.objects.property()
-    @Input
-    val apiUrl: Property<String> = project.objects.property()
-    @Input
-    val pkgName: Property<String> = project.objects.property()
-    @Input
-    val repo: Property<String> = project.objects.property()
-    @Input
-    @Optional
-    val userOrg: Property<String> = project.objects.property()
-    @Input
-    @Optional
-    val version: Property<String> = project.objects.property()
-    @Input
-    @Optional
-    val autoPublishWaitForSeconds: Property<Int> = project.objects.property()
-
-    private companion object {
-        const val UNSET = "UNSET"
-    }
+open class NebulaBintrayVersionTask : NebulaBintrayAbstractTask() {
 
     @TaskAction
     fun createVersion() {
@@ -66,6 +38,4 @@ open class NebulaBintrayVersionTask : DefaultTask() {
             throw GradleException("Received ${result.code()} attempting to publish $resolvedPkgName version $resolvedVersion")
         }
     }
-
-    private fun String.isNotSet() = this == UNSET
 }
